+++ .eslintrc.js
@@ -31,7 +31,7 @@ module.exports = {
         ecmaFeatures: {
             jsx: true,
         },
-        project: ["./tsconfig.json"],
+        project: ["./eslint/tsconfig.json"],
     },
     ignorePatterns: [".eslintrc.js", "jest.config.js"],
     rules: {
